AdminUserServiceRefactor

+++ src/main/java/softuniBlog/controller/admin/AdminUserController.java
@@ -15,26 +15,48 @@ public class AdminUserController extends AdminUserImpl {
 
     @GetMapping("/")
     public String listUsers(Model model){
-        return loadlistUsersView(model);
+        model.addAttribute("users", getUserRepository().findAll());
+        model.addAttribute("view", "admin/users/list");
+
+        return "base-layout";
     }
 
     @GetMapping("/edit/{id}")
     public String edit(@PathVariable Integer id, Model model){
-        return loadUserEditView(id, model);
+        if(!getUserRepository().existsById(id)){
+            return "redirect:/admin/users/";
+        }
+        model.addAttribute("user", getUserRepository().getOne(id));
+        model.addAttribute("roles", getRoleRepository().findAll());
+        model.addAttribute("view", "admin/users/edit");
+        return "base-layout";
     };
 
     @PostMapping("/edit/{id}")
     public String editProcess(@PathVariable Integer id, UserEditBindingModel userEditBindingModel){
-        return editUser(id, userEditBindingModel);
+        if(!getUserRepository().existsById(id)){
+            return "redirect:/admin/users/";
+        }
+        editUser(id, userEditBindingModel);
+        return "redirect:/admin/users/";
     }
 
     @GetMapping("/delete/{id}")
     public String delete(@PathVariable Integer id, Model model){
-        return loadUserDeleteView(id, model);
+        if(!getUserRepository().existsById(id)){
+            return "redirect:/admin/users";
+        }
+        model.addAttribute("user", getUserRepository().getOne(id));
+        model.addAttribute("view", "admin/users/delete");
+        return "base-layout";
     }
 
     @PostMapping("/delete/{id}")
     public String deleteProcess(@PathVariable Integer id){
-        return deleteUser(id);
+        if(!getUserRepository().existsById(id)){
+            return "redirect:/admin/users/";
+        }
+        deleteUser(id);
+        return "redirect:/admin/users/";
     }
 }
